+++ baybe/kernels/basic.py
@@ -6,14 +6,14 @@ from attrs.converters import optional as optional_c
 from attrs.validators import ge, gt, in_, instance_of
 from attrs.validators import optional as optional_v
 
-from baybe.kernels.base import Kernel
+from baybe.kernels.base import BasicKernel
 from baybe.priors.base import Prior
 from baybe.utils.conversion import fraction_to_float
 from baybe.utils.validation import finite_float
 
 
 @define(frozen=True)
-class LinearKernel(Kernel):
+class LinearKernel(BasicKernel):
     """A linear kernel."""
 
     variance_prior: Prior | None = field(
@@ -43,7 +43,7 @@ class LinearKernel(Kernel):
 
 
 @define(frozen=True)
-class MaternKernel(Kernel):
+class MaternKernel(BasicKernel):
     """A Matern kernel using a smoothness parameter."""
 
     nu: float = field(
@@ -68,7 +68,7 @@ class MaternKernel(Kernel):
 
 
 @define(frozen=True)
-class PeriodicKernel(Kernel):
+class PeriodicKernel(BasicKernel):
     """A periodic kernel."""
 
     lengthscale_prior: Prior | None = field(
@@ -112,7 +112,7 @@ class PeriodicKernel(Kernel):
 
 
 @define(frozen=True)
-class PiecewisePolynomialKernel(Kernel):
+class PiecewisePolynomialKernel(BasicKernel):
     """A piecewise polynomial kernel."""
 
     q: int = field(validator=in_([0, 1, 2, 3]), default=2)
@@ -132,7 +132,7 @@ class PiecewisePolynomialKernel(Kernel):
 
 
 @define(frozen=True)
-class PolynomialKernel(Kernel):
+class PolynomialKernel(BasicKernel):
     """A polynomial kernel."""
 
     power: int = field(validator=[instance_of(int), ge(0)])
@@ -163,7 +163,7 @@ class PolynomialKernel(Kernel):
 
 
 @define(frozen=True)
-class RBFKernel(Kernel):
+class RBFKernel(BasicKernel):
     """A radial basis function (RBF) kernel."""
 
     lengthscale_prior: Prior | None = field(
@@ -180,7 +180,7 @@ class RBFKernel(Kernel):
 
 
 @define(frozen=True)
-class RFFKernel(Kernel):
+class RFFKernel(BasicKernel):
     """A random Fourier features (RFF) kernel."""
 
     num_samples: int = field(validator=[instance_of(int), ge(1)])
@@ -200,7 +200,7 @@ class RFFKernel(Kernel):
 
 
 @define(frozen=True)
-class RQKernel(Kernel):
+class RQKernel(BasicKernel):
     """A rational quadratic (RQ) kernel."""
 
     lengthscale_prior: Prior | None = field(